Conflicts:
	sql/create.sql
	sql/delete.sql
	sql/drop.sql
	sql/insert.sql

+++ src/main/java/at/ac/tuwien/sepm/ws16/qse01/dao/impl/JDBCProfileDAO.java
@@ -136,7 +136,7 @@ public class JDBCProfileDAO implements ProfileDAO {
             stmt.setBoolean(4,profile.isGreenscreenEnabled());
             stmt.setBoolean(5,profile.isMobilEnabled());
             stmt.setString(6,profile.getWatermark());
-            stmt.setLong(7,profile.getId());
+            stmt.setInt(7,profile.getId());
             stmt.executeUpdate();
             int returnUpdateCount = stmt.executeUpdate();
